APPEALS-28105 Addressed Code Climate errors
